Two conflicts gone
