--- controllers/phone/index.js
@@ -1,4 +1,0 @@
-var newCall = require('./new-call');
-var redirectCall = require('./redirect-call');
-
-module.exports = Object.assign({}, newCall, redirectCall);
